Web Gabriel Luan e Gabriel Lima (2)/Codigos/4_Scripts/requirements.txt.txt
pandas
flask
selenium
matplotlib.pyplot
seaborn
plotly.express
matplotlib.pyplot
missingno
unidecode
numpy